+++ checkpoint/orbax/checkpoint/_src/handlers/base_pytree_checkpoint_handler.py
@@ -788,7 +788,8 @@ class BasePyTreeCheckpointHandler(
         pytree_metadata_options=self._pytree_metadata_options,
     )
 
-  def metadata(self, directory: epath.Path) -> Optional[PyTree]:
+
+  def metadata(self, directory: epath.Path) -> tree_metadata.TreeMetadata:
     """Returns tree metadata.
 
     The result will be a PyTree matching the structure of the saved checkpoint.
@@ -815,8 +816,12 @@ class BasePyTreeCheckpointHandler(
       tree containing metadata.
     """
     is_ocdbt_checkpoint = type_handlers.is_ocdbt_checkpoint(directory)
-    return self._read_metadata_file(directory).as_user_metadata(
-        directory, self._type_handler_registry, use_ocdbt=is_ocdbt_checkpoint
+    return tree_metadata.build_default_tree_metadata(
+        self._read_metadata_file(directory).as_user_metadata(
+            directory,
+            self._type_handler_registry,
+            use_ocdbt=is_ocdbt_checkpoint,
+        ),
     )
 
   def finalize(self, directory: epath.Path) -> None: