+++ checkpoint/orbax/checkpoint/_src/testing/test_tree_utils.py
@@ -47,6 +47,11 @@ class MuNu(NamedTuple):
   nu: np.ndarray | None
 
 
+class IntegerNamedTuple(NamedTuple):
+  x: int | None
+  y: int | None
+
+
 class EmptyNamedTuple(NamedTuple):
   pass